// brokerchain-1.0.0/src/library/storage-simple-coll/queue-collection-all-load/type.ts
export interface Input {}

export interface OutputOk {
    version: number;
    items: { namespace: string; key: string; metadata: { [key: string]: any }; items: { id: string; [key: string]: any }[] }[];
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}
// ...
export type Output = OutputOk;